refactor: unify Whitney triangle surface-mass kernels in shared whitney_face module
silvermuller.rs and lumped_port.rs carried semantically identical
copies of the Whitney face surface-mass triplet kernel plus its
supporting geometry (face_geometry, TRI_LOCAL_EDGES /
BARYCENTRIC_MIDPOINTS, edge_lookup, vec3 helpers). Extract a shared
whitney_face module holding the triplet core and helpers; both
callers now delegate. Pure refactor — no numerical changes, existing
cross-validation test retained.

Closes #208
